% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{analyzeVariants}
\alias{analyzeVariants}
\title{Analysis of transcript variants}
\usage{
analyzeVariants(object, maxnvariant = 20, cores = 1)
}
\arguments{
\item{object}{\code{SGFeatureCounts} object}

\item{maxnvariant}{If more than \code{maxnvariant} variants are
identified in an event, the gene is skipped, resulting in a warning.
Set to \code{NA} to include all genes.}

\item{cores}{Number of cores available for parallel processing}
}
\value{
A \code{TxVariantCounts} object
}
\description{
High-level function for the analysis of transcript variants from
splice graph features. Transcript variants are identified with
\code{\link{findTxVariants}}. Representative counts and estimated
variant frequencies are obtained with \code{\link{getTxVariantCounts}}.
}
\examples{
txvc <- analyzeVariants(sgfc)
}
\author{
Leonard Goldstein
}

