% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotateTxVariants}
\alias{annotateTxVariants}
\title{Annotate transcript variants in terms of canonical events}
\usage{
annotateTxVariants(variants)
}
\arguments{
\item{variants}{\code{TxVariants} object}
}
\value{
\code{variants} with added elementMetadata column
  \dQuote{variantType} indicating canonical event(s)
}
\description{
Annotate transcript variants in terms of canonical events.
}
\details{
The following events are considered:
\dQuote{SE} (skipped exon),
\dQuote{S2E} (two consecutive exons skipped),
\dQuote{RI} (retained intron),
\dQuote{MXE} (mutually exclusive exons),
\dQuote{A5SS} (alternative 5' splice site),
\dQuote{A3SS} (alternative 3' splice site),
\dQuote{AFE} (alternative first exon),
\dQuote{ALE} (alternative last exon),
\dQuote{AS} (alternative start other than \dQuote{AFE}) and
\dQuote{AE} (alternative end other than \dQuote{ALE}).

These are binary events, defined by two alternative variants.
A variant is annotated as a canonical event if it coincides with one
of the two variants in the canonical event, and there is at least one
variant in the same event that coincides with the second variant of the
canonical event.
}
\author{
Leonard Goldstein
}
\keyword{internal}

