% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertToSGFeatures}
\alias{convertToSGFeatures}
\title{Convert transcript features to splice graph features}
\usage{
convertToSGFeatures(x, coerce = FALSE)
}
\arguments{
\item{x}{\code{TxFeatures} object}

\item{coerce}{Logical indicating whether transcript features
should be coerced to splice graph features without disjoining
exons and omitting splice donor and acceptor sites}
}
\value{
An \code{SGFeatures} object
}
\description{
Convert transcript features, predicted from RNA-seq data or
extracted from transcript annotation, to splice graph features.
}
\details{
Splice junctions are unaltered. Exons are disjoined into
non-overlapping exon bins. Adjacent exon bins without a splice site
at the shared boundary are merged. All exon bins are assigned type
\dQuote{E}.

Entries for splice donor and acceptor sites (positions immediately
upstream and downstream of introns, respectively) are added.

In the returned \code{SGFeatures} object, column slots \code{splice5p}
and \code{splice3p} indicate whether compatibility with an exon bin
requires a fragment to be spliced at the 5' or 3' boundary, respectively.
\code{splice5p} (\code{splice3p}) is \code{TRUE} if the first
(last) position of the exon coincides with a splice acceptor (donor),
and it is not adjacent to a neighboring exon bin.

Each feature is assigned a unique feature and gene identifier,
stored in column slots \code{featureID} and \code{geneID},
respectively. The latter indicates features that belong to the
same gene, represented by a connected component in the splice graph.
}
\examples{
sgf <- convertToSGFeatures(txf)
}
\author{
Leonard Goldstein
}

