% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertToTxFeatures}
\alias{convertToTxFeatures}
\title{Convert to TxFeatures object}
\usage{
convertToTxFeatures(x)
}
\arguments{
\item{x}{\code{TxDb} object, or \code{GRangesList} of exons
grouped by transcripts}
}
\value{
A \code{TxFeatures} object
}
\description{
Convert a \code{TxDb} object or a \code{GRangesList} of exons
grouped by transcripts to a \code{TxFeatures} object.
}
\details{
In the returned \code{TxFeatures} object, column slot \code{txName}
is based on either \code{tx_name} in the \code{TxDb} object
or \code{names(x)} if \code{x} is a \code{GRangesList}.
}
\examples{
gr <- GRanges(c(1, 1), IRanges(c(1, 201), c(100, 300)), c("+", "+"))
grl <- split(gr, 1)
txf <- convertToTxFeatures(grl)
}
\author{
Leonard Goldstein
}

