% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlaps.R
\name{exonCompatible}
\alias{exonCompatible}
\title{Compatible fragment counts for exons}
\usage{
exonCompatible(exons, spliceL, spliceR, frag_exonic, frag_intron,
  counts = TRUE)
}
\arguments{
\item{exons}{\code{IRanges} of exons}

\item{spliceL}{Logical vector indicating whether LHS boundary is spliced}

\item{spliceR}{Logical vector indicating whether RHS boundary is spliced}

\item{frag_exonic}{\code{IRangesList} of exonic regions, one entry
per fragment}

\item{frag_intron}{\code{IRangesList} of introns, one entry per fragment}

\item{counts}{Logical indicating whether counts or indices of
compatible fragments should be returned}
}
\value{
Counts or list of indices of compatible fragments
}
\description{
Identify fragments compatible with exons.
}
\author{
Leonard Goldstein
}
\keyword{internal}

