% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterTerminalExons}
\alias{filterTerminalExons}
\title{Filter terminal exons that share splice sites with internal exons}
\usage{
filterTerminalExons(features, min_overhang = NA, return_index = FALSE)
}
\arguments{
\item{features}{\code{TxFeatures} object}

\item{min_overhang}{For terminal exons sharing a splice site with an
internal exon, minimum overhang required for terminal exon to be
included. Use \code{NA} to remove all terminal exons sharing a splice
site with an internal exon.}

\item{return_index}{Logical indicating whether indices of retained
features should be returned instead of filtered features}
}
\value{
\code{TxFeatures} object with filtered features,
  or indices of retained features if \code{return_index = TRUE}
}
\description{
Terminal exons that share their splice site with an internal exon are
filtered based on the overhang with respect to the internal exon.
}
\details{
\code{predictTxFeatures} predicts flanking terminal exons for each
identified splice junction. Thus splice junctions are guaranteed to
have flanking exons, even after filtering exons during merging with
\code{mergeTxFeatures}. However, many predicted terminal exons share
a splice site with predicted internal exons and are often contained
within them. Many of these predictions are unlikely to be real terminal
exons and are excluded before further analysis.
}
\examples{
txf_filtered <- filterTerminalExons(txf)
}
\author{
Leonard Goldstein
}

