% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlaps.R
\name{findOverlapsRanges}
\alias{findOverlapsRanges}
\title{Modified \code{findOverlaps} function for \code{IRanges},
  \code{IRangesList} objects}
\usage{
findOverlapsRanges(query, subject, type = "any")
}
\arguments{
\item{query}{\code{IRanges} or \code{IRangesList} object}

\item{subject}{\code{IRanges} or \code{IRangesList} object}

\item{type}{Passed to \code{findOverlaps}}
}
\value{
\code{Hits} object
}
\description{
Modified \code{findOverlaps} function for \code{IRanges},
\code{IRangesList} objects that behaves analogous to
\code{findOverlaps} for \code{GRanges}, \code{GRangesList} objects.
}
\author{
Leonard Goldstein
}
\keyword{internal}

