% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphs.R
\name{findSGVariants}
\alias{findSGVariants}
\title{Find transcript variants from splice graph}
\usage{
findSGVariants(features, maxnvariant = 20, annotate_events = TRUE,
  cores = 1)
}
\arguments{
\item{features}{\code{SGFeatures} object}

\item{maxnvariant}{If more than \code{maxnvariant} variants are
identified in an event, the gene is skipped, resulting in a warning.
Set to \code{NA} to include all genes.}

\item{annotate_events}{Logical indicating whether identified
transcript variants should be annotated in terms of canonical events.
For details see help page for \code{\link{annotateSGVariants}}.}

\item{cores}{Number of cores available for parallel processing}
}
\value{
A \code{SGVariants} object
}
\description{
Find transcript variants from splice graph
}
\examples{
sgv <- findSGVariants(sgf)
}
\author{
Leonard Goldstein
}

