% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findTxVariants}
\alias{findTxVariants}
\title{Find transcript variants from splice graph}
\usage{
findTxVariants(features, maxnvariant = 20, annotate_events = TRUE,
  cores = 1)
}
\arguments{
\item{features}{\code{SGFeatures} object}

\item{maxnvariant}{If more than \code{maxnvariant} variants are
identified in an event, the gene is skipped, resulting in a warning.
Set to \code{NA} to include all genes.}

\item{annotate_events}{Logical indicating whether identified
transcript variants should be annotated in terms of canonical events.
For details see help page for \code{\link{annotateTxVariants}}.}

\item{cores}{Number of cores available for parallel processing}
}
\value{
A \code{TxVariants} object
}
\description{
Find transcript variants from splice graph
}
\examples{
txv <- findTxVariants(sgf)
}
\author{
Leonard Goldstein
}

