% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/counts.R
\name{getSGVariantCounts}
\alias{getSGVariantCounts}
\title{Representative counts and frequency estimates for
  transcript variants}
\usage{
getSGVariantCounts(object, variants, cores = 1)
}
\arguments{
\item{object}{\code{SGFeatureCounts} object}

\item{variants}{\code{SGVariants} object}

\item{cores}{Number of cores available for parallel processing}
}
\value{
A \code{SGVariantCounts} object
}
\description{
For transcript variants, obtain counts of compatible fragments
extending across the start and/or end of each variant.
Variant frequencies are estimated based on representive counts.
}
\examples{
sgvc <- getSGVariantCounts(sgfc, sgv)
}
\author{
Leonard Goldstein
}

