% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/counts.R
\name{getTxVariantCounts}
\alias{getTxVariantCounts}
\title{Representative counts and frequency estimates for
  transcript variants}
\usage{
getTxVariantCounts(object, variants)
}
\arguments{
\item{object}{\code{SGFeatureCounts} object}

\item{variants}{\code{TxVariants} object}
}
\value{
A \code{TxVariantCounts} object
}
\description{
For transcript variants, obtain counts of compatible fragments
extending across the start and/or end of each variant.
Variant frequencies are estimated based on representive counts.
}
\examples{
txvc <- getTxVariantCounts(sgfc, txv)
}
\author{
Leonard Goldstein
}

