% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{gr}
\alias{gr}
\title{Example region of interest}
\format{A \code{GRanges} object.}
\description{
FBXO31 gene locus, based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

