% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlaps.R
\name{junctionCompatible}
\alias{junctionCompatible}
\title{Compatible fragment counts for splice junctions}
\usage{
junctionCompatible(junctions, frag_intron, counts = TRUE)
}
\arguments{
\item{junctions}{\code{IRanges} of splice junctions}

\item{frag_intron}{\code{IRangesList} of introns, one entry per fragment}

\item{counts}{Logical indicating whether counts or indices of
compatible fragments should be returned}
}
\value{
Counts or list of indices of compatible fragments
}
\description{
Identify fragments compatible with splice junctions.
}
\author{
Leonard Goldstein
}
\keyword{internal}

