% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeSGFeatureCounts}
\alias{makeSGFeatureCounts}
\title{Create \code{SGFeatureCounts} object}
\usage{
makeSGFeatureCounts(rowData, colData, counts)
}
\arguments{
\item{rowData}{An \code{SGFeatures} object}

\item{colData}{Data frame with sample information}

\item{counts}{Integer matrix of counts}
}
\value{
An \code{SGFeatureCounts} object
}
\description{
Create \code{SGFeatureCounts} object from rowData, colData and counts.
}
\examples{
sgfc <- makeSGFeatureCounts(sgf, si, matrix(0L, length(sgf), nrow(si)))
}
\author{
Leonard Goldstein
}

