% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeVariantNames}
\alias{makeVariantNames}
\title{Create interpretable transcript variant names}
\usage{
makeVariantNames(variants)
}
\arguments{
\item{variants}{\code{TxVariants} object}
}
\value{
Character vector with transcript variant names
}
\description{
This function creates interpretable transcript variant names
taking the format GENE_EVENT_VARIANT/ORDER_TYPE.
GENE is based on geneName if available, and geneID otherwise.
EVENT and VARIANT enumerate events and variants for the same gene
and event, respectively. ORDER indicates the total number of
variants in the same event (e.g. 1/2 refers to the first out of two
transcript variants in the event). TYPE is based on variantType.
}
\examples{
makeVariantNames(txv)
}
\author{
Leonard Goldstein
}

