% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-merge.R
\name{mergeTxFeatures}
\alias{mergeTxFeatures}
\title{Merge redundant features}
\usage{
mergeTxFeatures(..., min_n_sample = 1)
}
\arguments{
\item{...}{one or more \code{TxFeatures} objects, or a single
list of \code{TxFeatures} objects}

\item{min_n_sample}{Minimum number of samples a feature must be
observed in to be included}
}
\value{
\code{TxFeatures} object with merged features
}
\description{
Merge features, typically after feature prediction in multiple samples.
}
\details{
Merged features are the union of splice junctions and internal exons.
For terminal exons with shared spliced boundary, the longest exon is
retained.
}
\examples{
txf_merged <- mergeTxFeatures(txf, txf)
}
\author{
Leonard Goldstein
}

