% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpliceGraph}
\alias{plotSpliceGraph}
\title{Plot splice graph}
\usage{
plotSpliceGraph(x, geneID = NULL, geneName = NULL, eventID = NULL,
  which = NULL, toscale = c("exon", "none", "gene"), label = c("id",
  "name", "label", "none"), color = "grey", color_novel = "red",
  color_alpha = 0.8, color_labels = FALSE, border = "fill", cexLab = 1,
  cexExon = 1, track = NULL, track_color = "darkblue",
  track_ylim = NULL, track_ypos = c(0.2, 0.1), track_nbin = 400,
  track_summary = mean, main = NULL, cexMain = 1, tx_view = FALSE,
  tx_dist = 0.2, tx_cex = 1, asp = 1)
}
\arguments{
\item{x}{\code{SGFeatures} or \code{TxVariants} object}

\item{geneID}{Single gene identifier used to subset \code{x}}

\item{geneName}{Single gene name used to subset \code{x}}

\item{eventID}{Single event identifier used to subset \code{x}}

\item{which}{\code{GRanges} used to subset \code{x}}

\item{toscale}{Controls which parts of the splice graph are drawn to
scale. Possible values are \dQuote{none} (exonic and intronic regions
have constant length), \dQuote{exon} (exonic regions are drawn to scale)
and \dQuote{gene} (both exonic and intronic regions are drawn to scale).}

\item{label}{Format of exon/splice junction labels,
possible values are \dQuote{id} (format E1,... J1,...), \dQuote{name}
(format type:chromosome:start-end:strand), \dQuote{label} for labels
specified in elementMetadata column \dQuote{label}, or \dQuote{none}
for no labels.}

\item{color}{Color used for plotting the splice graph. Ignored if features
elementMetadata column \dQuote{color} is not \code{NULL}.}

\item{color_novel}{Features with missing annotation are
highlighted in \code{color_novel}. Ignored if features
elementMetadata column \dQuote{color} is not \code{NULL}.}

\item{color_alpha}{Controls color transparency}

\item{color_labels}{Logical indicating whether label colors should
be the same as feature colors}

\item{border}{Determines the color of exon borders, can be \dQuote{fill}
(same as exon color), \dQuote{none} (no border) or a valid color name}

\item{cexLab}{Scale factor for feature labels}

\item{cexExon}{Scale factor for exon height}

\item{track}{\code{RLeList} containing nucleotide-level scores or a
\code{GRangesList} to be plotted with the splice graph}

\item{track_color}{Color used for plotting tracks}

\item{track_ylim}{y-axis range used for plotting scores}

\item{track_ypos}{Numeric vector of length two, indicating the vertical
position and height of the track panel, specificed as fractions of the
height of the plotting region}

\item{track_nbin}{Number of bins for plotting scores}

\item{track_summary}{Function used to calculate per-bin score summaries}

\item{main}{Plot title}

\item{cexMain}{Scale factor for plot title}

\item{tx_view}{Plot transcripts instead of splice graph (experimental)}

\item{tx_dist}{Vertical distance between transcripts as fraction of height
of plotting region}

\item{tx_cex}{Scale factor for transcript labels}

\item{asp}{Aspect ratio of graphics region}
}
\value{
\code{data.frame} with plotting information for exons and
  splice junctions in the splice graph
}
\description{
Plot splice graph implied by splice junctions and exon bins.
}
\details{
By default, splice graph feature color is determined by annotation
(see arguments \code{color}, \code{color_novel}) and labels are
generated automatically (see argument \code{label}).

Alternatively, colors and labels can be specified via elementMetadata
columns \dQuote{color} and \dQuote{label}, respectively.
}
\examples{
\dontrun{
  sgf_annotated <- annotate(sgf, txf)
  plotSpliceGraph(sgf_annotated)
}
\dontrun{
  txv_annotated <- annotate(txv, txf)
  plotSpliceGraph(txv_annotated)
}
}
\author{
Leonard Goldstein
}

