% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-prediction.R
\name{predictCandidatesInternal}
\alias{predictCandidatesInternal}
\title{Identify candidate internal exons}
\usage{
predictCandidatesInternal(islands, splicesites, frag_coverage, relCov)
}
\arguments{
\item{islands}{\code{IRanges} of genomic regions with minimal read
coverage required for internal exon prediction}

\item{splicesites}{\code{IRanges} of splice sites with metadata
columns \dQuote{type} and \dQuote{N}}

\item{frag_coverage}{\code{Rle} object with fragment coverage}

\item{relCov}{Minimum relative coverage required for exon prediction}
}
\value{
\code{IRanges} of candidate internal exons
}
\description{
Identify candidate internal exons based on previously identified
splice sites and regions with sufficient read coverage.
}
\author{
Leonard Goldstein
}
\keyword{internal}

