% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictTxFeatures}
\alias{predictTxFeatures}
\title{Transcript feature prediction from BAM files}
\usage{
predictTxFeatures(sample_info, which = NULL, alpha = 2, psi = 0,
  beta = 0.2, gamma = 0.2, min_junction_count = NULL, min_n_sample = 1,
  min_overhang = NA, cores_per_sample = 1, BPPARAM = MulticoreParam(1))
}
\arguments{
\item{sample_info}{\code{data.frame} containing sample information.
Columns \dQuote{sample_name}, \dQuote{file_bam} and \dQuote{paired_end}
are sufficient for use with argument \code{min_junction_count}.
For use with argument \code{alpha}, which specifies minimum fragment
count for splice junction prediction in FPKM units, additional
columns \dQuote{read_length}, \dQuote{frag_length} and \dQuote{lib_size}
are required.}

\item{which}{\code{GRanges} of genomic regions to be considered for
feature prediction, passed to \code{ScanBamParam}}

\item{alpha}{Minimum FPKM required for a splice junction to be
included. Internally, FPKMs are converted to counts, requiring arguments
\code{read_length}, \code{frag_length} and \code{lib_size}.
\code{alpha} is ignored if argument \code{min_junction_count}
is specified.}

\item{psi}{Minimum splice frequency required for a splice junction
to be included}

\item{beta}{Minimum relative coverage required for an internal exon
to be included}

\item{gamma}{Minimum relative coverage required for a terminal exon
to be included}

\item{min_junction_count}{Minimum fragment count required for a splice
junction to be included. If specified, argument \code{alpha} is ignored.}

\item{min_n_sample}{Minimum number of samples a feature must be
observed in to be included}

\item{min_overhang}{After merging, terminal exons are processed.
For terminal exons sharing a splice site with an internal exon,
minimum overhang required for terminal exons to be included.
For remaining terminal exons overlapping other exons, minimum
overhang required to suppress trimming. Use \code{NA} to remove all
terminal exons sharing a splice with an internal exon and trim all
remaining terminal exons overlapping other exons. Use \code{NULL}
to disable processing (disabling processing is useful if results are
subsequently merged with other predictions and processing is
postponed until after the merging step).}

\item{cores_per_sample}{Number of cores per sample}

\item{BPPARAM}{\code{BiocParallelParam} for processing samples in parallel,
defaults to \code{MulticoreParam(1)}}
}
\value{
A \code{TxFeatures} object
}
\description{
Transcript features are predicted for each sample and merged across
samples. Subsequently, terminal exons are filtered and trimmed
(if applicable). For details, see the help pages for
\code{\link{predictTxFeaturesPerSample}}, \code{\link{mergeTxFeatures}},
and \code{\link{processTerminalExons}}.
}
\examples{
dir <- system.file("extdata", package = "SGSeq")
si$file_bam <- file.path(dir, "bams", si$file_bam)
txf <- predictTxFeatures(si, gr)
}
\author{
Leonard Goldstein
}

