% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictTxFeaturesRanges}
\alias{predictTxFeaturesRanges}
\title{Identification of splice junctions and exons for a given
  chromosome and strand}
\usage{
predictTxFeaturesRanges(file_bam, paired_end, which, min_junction_count, psi,
  beta, gamma, include_counts, retain_coverage, junctions_only)
}
\arguments{
\item{file_bam}{BAM file with genomic RNA-seq read alignments}

\item{paired_end}{Logical, \code{TRUE} for paired-end data,
\code{FALSE} for single-end data}

\item{which}{\code{GRanges} of genomic regions to be considered for
feature prediction, passed to \code{ScanBamParam}}

\item{min_junction_count}{Minimum fragment count required for a splice
junction to be included. If specified, argument \code{alpha} is ignored.}

\item{psi}{Minimum splice frequency required for a splice junction
to be included}

\item{beta}{Minimum relative coverage required for an internal exon
to be included}

\item{gamma}{Minimum relative coverage required for a terminal exon
to be included}

\item{include_counts}{Logical indicating whether counts of
compatible fragments should be included in elementMetadata column
\dQuote{N}}

\item{retain_coverage}{Logical indicating whether coverages for each
exon should be retained as an \code{IntegerList} in elementMetadata
column \dQuote{coverage}. This allows filtering of features
using more stringent criteria after the initial prediction.}

\item{junctions_only}{Logical indicating whether predictions
should be limited to identification of splice junctions only}
}
\value{
\code{GRanges} of predicted features
}
\description{
Identification of splice junctions and exons for a given
  chromosome and strand
}
\author{
Leonard Goldstein
}
\keyword{internal}

