% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{sgf}
\alias{sgf}
\title{Example splice graph features}
\format{An \code{SGFeatures} object.}
\description{
Splice graph features for FBXO31, predicted from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

