% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgf_ann}
\alias{sgf_ann}
\title{Example splice graph features (annotation-based)}
\format{An \code{SGFeatures} object.}
\description{
Splice graph features for FBXO31, based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

