% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgf_pred}
\alias{sgf_pred}
\title{Example splice graph features (predicted)}
\format{An \code{SGFeatures} object.}
\description{
Splice graph features for FBXO31, predicted from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

