% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgfc_ann}
\alias{sgfc_ann}
\title{Example splice graph feature counts (annotation-based)}
\format{An \code{SGFeatureCounts} object.}
\description{
Compatible counts and FPKMs for FBXO31 splice graph features,
based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

