% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgfc_pred}
\alias{sgfc_pred}
\title{Example splice graph feature counts (predicted)}
\format{An \code{SGFeatureCounts} object.}
\description{
Compatible counts and FPKMs for FBXO31 splice graph features,
predicted from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

