% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\name{SGSegments}
\alias{SGSegments}
\title{Constructor function for S4 class \code{SGSegments}}
\usage{
SGSegments(x)
}
\arguments{
\item{x}{\code{GRangesList} of \code{SGFeatures} with appropriate
outer elementMetadata columns}
}
\value{
A \code{SGSegments} object
}
\description{
Creates an instance of S4 class \code{SGSegments} for storing
splice graph segments.
}
\author{
Leonard Goldstein
}
\keyword{internal}

