% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgv_ann}
\alias{sgv_ann}
\title{Example splice variants (annotation-based)}
\format{A \code{SGVariants} object.}
\description{
Splice variants for FBXO31, based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

