% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgv_pred}
\alias{sgv_pred}
\title{Example splice variants (predicted)}
\format{A \code{SGVariants} object.}
\description{
Splice variants for FBXO31, predicted from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

