% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SGVariantCounts}
\alias{SGVariantCounts}
\title{Splice graph variant counts}
\usage{
SGVariantCounts(x)
}
\arguments{
\item{x}{\code{RangedSummarizedExperiment} with \code{SGVariants}
as \code{rowRanges} and assays \dQuote{variantFreq},
\dQuote{countsVariant5p}, \dQuote{countsVariant3p},
\dQuote{countsEvent5p}, \dQuote{countsEvent3p}, and
optionally \dQuote{countsVariant5pOr3p}}
}
\value{
\code{SGVariantCounts} object
}
\description{
Creates an instance of S4 class \code{SGVariantCounts} for storing
splice variant counts.
}
\examples{
sgvc <- SGVariantCounts()
}
\author{
Leonard Goldstein
}
