% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgvc_ann}
\alias{sgvc_ann}
\title{Example splice variant counts (annotated)}
\format{A \code{SGVariantCounts} object.}
\description{
Splice variant counts and frequencies for FBXO31.
Splice variants are based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

