% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgvc_ann_from_bam}
\alias{sgvc_ann_from_bam}
\title{Example splice variant counts (annotated) from BAM files}
\format{A \code{SGVariantCounts} object.}
\description{
Splice variant counts and frequencies for FBXO31.
Splice variants are based on UCSC knownGene annotation.
Counts were obtained from BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

