% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sgvc_pred}
\alias{sgvc_pred}
\title{Example splice variant counts (predicted)}
\format{A \code{SGVariantCounts} object.}
\description{
Splice variant counts and frequencies for FBXO31.
Splice variants were predicted from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

