% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{si}
\alias{si}
\title{Example sample information}
\format{A \code{data.frame} with columns \dQuote{sample_name},
\dQuote{file_bam}, \dQuote{paired_end}, \dQuote{read_length},
\dQuote{frag_length} and \dQuote{lib_size}.}
\description{
Sample information for example BAM files included in the
\code{SGSeq} package.
}
\author{
Leonard Goldstein
}
\keyword{internal}

