% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slots-generics.R, R/slots-methods.R
\docType{methods}
\name{slots}
\alias{closed3p}
\alias{closed3p,SGVariantCounts-method}
\alias{closed3p,SGVariants-method}
\alias{closed3p<-}
\alias{closed3p<-,SGVariantCounts-method}
\alias{closed3p<-,SGVariants-method}
\alias{closed3pEvent}
\alias{closed3pEvent,SGVariantCounts-method}
\alias{closed3pEvent,SGVariants-method}
\alias{closed3pEvent<-}
\alias{closed3pEvent<-,SGVariantCounts-method}
\alias{closed3pEvent<-,SGVariants-method}
\alias{closed5p}
\alias{closed5p,SGVariantCounts-method}
\alias{closed5p,SGVariants-method}
\alias{closed5p<-}
\alias{closed5p<-,SGVariantCounts-method}
\alias{closed5p<-,SGVariants-method}
\alias{closed5pEvent}
\alias{closed5pEvent,SGVariantCounts-method}
\alias{closed5pEvent,SGVariants-method}
\alias{closed5pEvent<-}
\alias{closed5pEvent<-,SGVariantCounts-method}
\alias{closed5pEvent<-,SGVariants-method}
\alias{eventID}
\alias{eventID,SGVariantCounts-method}
\alias{eventID,SGVariants-method}
\alias{eventID<-}
\alias{eventID<-,SGVariantCounts-method}
\alias{eventID<-,SGVariants-method}
\alias{featureID}
\alias{featureID,Counts-method}
\alias{featureID,Paths-method}
\alias{featureID,SGFeatures-method}
\alias{featureID3p}
\alias{featureID3p,SGVariantCounts-method}
\alias{featureID3p,SGVariants-method}
\alias{featureID3p<-}
\alias{featureID3p<-,SGVariantCounts-method}
\alias{featureID3p<-,SGVariants-method}
\alias{featureID3pEvent}
\alias{featureID3pEvent,SGVariantCounts-method}
\alias{featureID3pEvent,SGVariants-method}
\alias{featureID3pEvent<-}
\alias{featureID3pEvent<-,SGVariantCounts-method}
\alias{featureID3pEvent<-,SGVariants-method}
\alias{featureID5p}
\alias{featureID5p,SGVariantCounts-method}
\alias{featureID5p,SGVariants-method}
\alias{featureID5p<-}
\alias{featureID5p<-,SGVariantCounts-method}
\alias{featureID5p<-,SGVariants-method}
\alias{featureID5pEvent}
\alias{featureID5pEvent,SGVariantCounts-method}
\alias{featureID5pEvent,SGVariants-method}
\alias{featureID5pEvent<-}
\alias{featureID5pEvent<-,SGVariantCounts-method}
\alias{featureID5pEvent<-,SGVariants-method}
\alias{featureID<-}
\alias{featureID<-,Counts-method}
\alias{featureID<-,Paths-method}
\alias{featureID<-,SGFeatures-method}
\alias{from,Paths-method}
\alias{from,SGVariantCounts-method}
\alias{from<-}
\alias{from<-,Paths-method}
\alias{from<-,SGVariantCounts-method}
\alias{geneID}
\alias{geneID,Counts-method}
\alias{geneID,Paths-method}
\alias{geneID,SGFeatures-method}
\alias{geneID<-}
\alias{geneID<-,Counts-method}
\alias{geneID<-,Paths-method}
\alias{geneID<-,SGFeatures-method}
\alias{geneName}
\alias{geneName,Counts-method}
\alias{geneName,Features-method}
\alias{geneName,Paths-method}
\alias{geneName<-}
\alias{geneName<-,Counts-method}
\alias{geneName<-,Features-method}
\alias{geneName<-,Paths-method}
\alias{segmentID}
\alias{segmentID,Paths-method}
\alias{segmentID,SGVariantCounts-method}
\alias{segmentID<-}
\alias{segmentID<-,Paths-method}
\alias{segmentID<-,SGVariantCounts-method}
\alias{slots}
\alias{splice3p}
\alias{splice3p,SGFeatureCounts-method}
\alias{splice3p,SGFeatures-method}
\alias{splice3p,SGSegments-method}
\alias{splice3p<-}
\alias{splice3p<-,SGFeatureCounts-method}
\alias{splice3p<-,SGFeatures-method}
\alias{splice3p<-,SGSegments-method}
\alias{splice5p}
\alias{splice5p,SGFeatureCounts-method}
\alias{splice5p,SGFeatures-method}
\alias{splice5p,SGSegments-method}
\alias{splice5p<-}
\alias{splice5p<-,SGFeatureCounts-method}
\alias{splice5p<-,SGFeatures-method}
\alias{splice5p<-,SGSegments-method}
\alias{to,Paths-method}
\alias{to,SGVariantCounts-method}
\alias{to<-}
\alias{to<-,Paths-method}
\alias{to<-,SGVariantCounts-method}
\alias{txName}
\alias{txName,Counts-method}
\alias{txName,Features-method}
\alias{txName,Paths-method}
\alias{txName<-}
\alias{txName<-,Counts-method}
\alias{txName<-,Features-method}
\alias{txName<-,Paths-method}
\alias{type,Counts-method}
\alias{type,Features-method}
\alias{type,Paths-method}
\alias{type<-}
\alias{type<-,Counts-method}
\alias{type<-,Features-method}
\alias{type<-,Paths-method}
\alias{variantID}
\alias{variantID,SGVariantCounts-method}
\alias{variantID,SGVariants-method}
\alias{variantID<-}
\alias{variantID<-,SGVariantCounts-method}
\alias{variantID<-,SGVariants-method}
\alias{variantName}
\alias{variantName,SGVariantCounts-method}
\alias{variantName,SGVariants-method}
\alias{variantName<-}
\alias{variantName<-,SGVariantCounts-method}
\alias{variantName<-,SGVariants-method}
\alias{variantType}
\alias{variantType,SGVariantCounts-method}
\alias{variantType,SGVariants-method}
\alias{variantType<-}
\alias{variantType<-,SGVariantCounts-method}
\alias{variantType<-,SGVariants-method}
\title{Accessing and replacing metadata columns}
\usage{
type(x) <- value

txName(x)

txName(x) <- value

geneName(x)

geneName(x) <- value

featureID(x)

featureID(x) <- value

geneID(x)

geneID(x) <- value

splice5p(x)

splice5p(x) <- value

splice3p(x)

splice3p(x) <- value

from(x) <- value

to(x) <- value

segmentID(x)

segmentID(x) <- value

variantID(x)

variantID(x) <- value

eventID(x)

eventID(x) <- value

closed5p(x)

closed5p(x) <- value

closed3p(x)

closed3p(x) <- value

closed5pEvent(x)

closed5pEvent(x) <- value

closed3pEvent(x)

closed3pEvent(x) <- value

variantType(x)

variantType(x) <- value

variantName(x)

variantName(x) <- value

featureID5p(x)

featureID5p(x) <- value

featureID3p(x)

featureID3p(x) <- value

featureID5pEvent(x)

featureID5pEvent(x) <- value

featureID3pEvent(x)

featureID3pEvent(x) <- value

\S4method{type}{Features}(x)

\S4method{type}{Paths}(x)

\S4method{type}{Counts}(x)

\S4method{type}{Features}(x) <- value

\S4method{type}{Paths}(x) <- value

\S4method{type}{Counts}(x) <- value

\S4method{txName}{Features}(x)

\S4method{txName}{Paths}(x)

\S4method{txName}{Counts}(x)

\S4method{txName}{Features}(x) <- value

\S4method{txName}{Paths}(x) <- value

\S4method{txName}{Counts}(x) <- value

\S4method{geneName}{Features}(x)

\S4method{geneName}{Paths}(x)

\S4method{geneName}{Counts}(x)

\S4method{geneName}{Features}(x) <- value

\S4method{geneName}{Paths}(x) <- value

\S4method{geneName}{Counts}(x) <- value

\S4method{featureID}{SGFeatures}(x)

\S4method{featureID}{Paths}(x)

\S4method{featureID}{Counts}(x)

\S4method{featureID}{SGFeatures}(x) <- value

\S4method{featureID}{Paths}(x) <- value

\S4method{featureID}{Counts}(x) <- value

\S4method{geneID}{SGFeatures}(x)

\S4method{geneID}{Paths}(x)

\S4method{geneID}{Counts}(x)

\S4method{geneID}{SGFeatures}(x) <- value

\S4method{geneID}{Paths}(x) <- value

\S4method{geneID}{Counts}(x) <- value

\S4method{splice5p}{SGFeatures}(x)

\S4method{splice5p}{SGSegments}(x)

\S4method{splice5p}{SGFeatureCounts}(x)

\S4method{splice5p}{SGFeatures}(x) <- value

\S4method{splice5p}{SGSegments}(x) <- value

\S4method{splice5p}{SGFeatureCounts}(x) <- value

\S4method{splice3p}{SGFeatures}(x)

\S4method{splice3p}{SGSegments}(x)

\S4method{splice3p}{SGFeatureCounts}(x)

\S4method{splice3p}{SGFeatures}(x) <- value

\S4method{splice3p}{SGSegments}(x) <- value

\S4method{splice3p}{SGFeatureCounts}(x) <- value

\S4method{segmentID}{Paths}(x)

\S4method{segmentID}{SGVariantCounts}(x)

\S4method{segmentID}{Paths}(x) <- value

\S4method{segmentID}{SGVariantCounts}(x) <- value

\S4method{from}{Paths}(x)

\S4method{from}{SGVariantCounts}(x)

\S4method{from}{Paths}(x) <- value

\S4method{from}{SGVariantCounts}(x) <- value

\S4method{to}{Paths}(x)

\S4method{to}{SGVariantCounts}(x)

\S4method{to}{Paths}(x) <- value

\S4method{to}{SGVariantCounts}(x) <- value

\S4method{eventID}{SGVariants}(x)

\S4method{eventID}{SGVariantCounts}(x)

\S4method{eventID}{SGVariants}(x) <- value

\S4method{eventID}{SGVariantCounts}(x) <- value

\S4method{variantID}{SGVariants}(x)

\S4method{variantID}{SGVariantCounts}(x)

\S4method{variantID}{SGVariants}(x) <- value

\S4method{variantID}{SGVariantCounts}(x) <- value

\S4method{closed5p}{SGVariants}(x)

\S4method{closed5p}{SGVariantCounts}(x)

\S4method{closed5p}{SGVariants}(x) <- value

\S4method{closed5p}{SGVariantCounts}(x) <- value

\S4method{closed3p}{SGVariants}(x)

\S4method{closed3p}{SGVariantCounts}(x)

\S4method{closed3p}{SGVariants}(x) <- value

\S4method{closed3p}{SGVariantCounts}(x) <- value

\S4method{closed5pEvent}{SGVariants}(x)

\S4method{closed5pEvent}{SGVariantCounts}(x)

\S4method{closed5pEvent}{SGVariants}(x) <- value

\S4method{closed5pEvent}{SGVariantCounts}(x) <- value

\S4method{closed3pEvent}{SGVariants}(x)

\S4method{closed3pEvent}{SGVariantCounts}(x)

\S4method{closed3pEvent}{SGVariants}(x) <- value

\S4method{closed3pEvent}{SGVariantCounts}(x) <- value

\S4method{variantName}{SGVariants}(x)

\S4method{variantName}{SGVariantCounts}(x)

\S4method{variantName}{SGVariants}(x) <- value

\S4method{variantName}{SGVariantCounts}(x) <- value

\S4method{variantType}{SGVariants}(x)

\S4method{variantType}{SGVariantCounts}(x)

\S4method{variantType}{SGVariants}(x) <- value

\S4method{variantType}{SGVariantCounts}(x) <- value

\S4method{featureID5p}{SGVariants}(x)

\S4method{featureID5p}{SGVariantCounts}(x)

\S4method{featureID5p}{SGVariants}(x) <- value

\S4method{featureID5p}{SGVariantCounts}(x) <- value

\S4method{featureID3p}{SGVariants}(x)

\S4method{featureID3p}{SGVariantCounts}(x)

\S4method{featureID3p}{SGVariants}(x) <- value

\S4method{featureID3p}{SGVariantCounts}(x) <- value

\S4method{featureID5pEvent}{SGVariants}(x)

\S4method{featureID5pEvent}{SGVariantCounts}(x)

\S4method{featureID5pEvent}{SGVariants}(x) <- value

\S4method{featureID5pEvent}{SGVariantCounts}(x) <- value

\S4method{featureID3pEvent}{SGVariants}(x)

\S4method{featureID3pEvent}{SGVariantCounts}(x)

\S4method{featureID3pEvent}{SGVariants}(x) <- value

\S4method{featureID3pEvent}{SGVariantCounts}(x) <- value
}
\arguments{
\item{x}{Object containing metadata column}

\item{value}{Replacement value}
}
\value{
Content of metadata column for accessor functions or updated
  object for replacement functions.
}
\description{
Accessor and replacement functions for metadata columns.
}
\details{
S4 classes defined in the \code{SGSeq} package contain metadata columns
that store information for each element in the object. For example, class
\code{TxFeatures} contains a column \code{type} that indicates feature
type. The specific columns contained in an object depend on its class.
}
\examples{
head(type(txf_ann))
head(type(sgf_ann))
}
\author{
Leonard Goldstein
}

