% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{tx}
\alias{tx}
\title{Example transcripts}
\format{A \code{GRangesList} object.}
\description{
FBXO31 transcripts, based on UCSC knownGene annotation.
Suitable as input for \code{convertToTxFeatures} and
\code{predictVariantConsequences}.
}
\author{
Leonard Goldstein
}
\keyword{internal}

