% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{txf}
\alias{txf}
\title{Example transcript features}
\format{A \code{TxFeatures} object.}
\description{
Transcript features for FBXO31, based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

