% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{txf_ann}
\alias{txf_ann}
\title{Example transcript features (annotation-based)}
\format{A \code{TxFeatures} object.}
\description{
Transcript features for FBXO31, based on UCSC knownGene annotation.
}
\author{
Leonard Goldstein
}
\keyword{internal}

