% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{txf_pred}
\alias{txf_pred}
\title{Example transcript features (predicted)}
\format{A \code{TxFeatures} object.}
\description{
Transcript features for FBXO31, predicted from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

