% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\name{TxVariantCounts}
\alias{TxVariantCounts}
\title{Constructor function for S4 class \code{SGFeatureCounts}}
\usage{
TxVariantCounts(x)
}
\arguments{
\item{x}{\code{SummarizedExperiment} with \code{TxVariants}
as \code{rowData} and appropriate assays}
}
\value{
A \code{TxVariantCounts} object
}
\description{
Creates an instance of S4 class \code{TxVariantCounts} for storing
representative transcript variant counts.
}
\examples{
txvc <- TxVariantCounts()
}
\author{
Leonard Goldstein
}

