% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\name{TxVariants}
\alias{TxVariants}
\title{Constructor function for S4 class \code{TxVariants}}
\usage{
TxVariants(x)
}
\arguments{
\item{x}{\code{GRangesList} of \code{SGFeatures} with appropriate
outer elementMetadata columns}
}
\value{
A \code{TxVariants} object
}
\description{
Creates an instance of S4 class \code{TxVariants} for storing
transcript variants.
}
\examples{
txv <- TxVariants()
}
\author{
Leonard Goldstein
}

