% Generated by roxygen2 (4.0.2): do not edit by hand
\name{txvc}
\alias{txvc}
\title{Example transcript variant counts}
\format{A \code{TxVariantCounts} object.}
\description{
Transcript variant counts and frequencies for FBXO31. Splice graph
features were predicted and counts obtained from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

