% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,SGVariants-method}
\alias{updateObject,SGVariantCounts-method}
\title{Update object}
\usage{
\S4method{updateObject}{SGVariants}(object, ..., verbose = FALSE)

\S4method{updateObject}{SGVariantCounts}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object to be updated}

\item{...}{Additional arguments}

\item{verbose}{Should a warning message be generated}
}
\value{
Updated object
}
\description{
Update object created with previous version of SGSeq.
}
\author{
Leonard Goldstein
}
