% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{analyzeVariants}
\alias{analyzeVariants}
\title{Analysis of splice variants}
\usage{
analyzeVariants(object, maxnvariant = 20, cores = 1)
}
\arguments{
\item{object}{\code{SGFeatureCounts} object}

\item{maxnvariant}{If more than \code{maxnvariant} variants are
identified in an event, the gene is skipped, resulting in a warning.
Set to \code{NA} to include all genes.}

\item{cores}{Number of cores available for parallel processing}
}
\value{
An \code{SGVariantCounts} object
}
\description{
High-level function for the analysis of splice variants from
splice graph features. Splice variants are identified with
\code{\link{findSGVariants}}. Representative counts are obtained
and variant frequencies estimated with \code{\link{getSGVariantCounts}}.
}
\examples{
sgvc <- analyzeVariants(sgfc)
}
\author{
Leonard Goldstein
}

