% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate}
\alias{annotate}
\title{Annotation with respect to transcript features}
\usage{
annotate(query, subject)
}
\arguments{
\item{query}{\code{SGFeatures}, \code{SGVariants},
\code{SGFeatureCounts} or \code{SGVariantCounts} object}

\item{subject}{\code{TxFeatures} object}
}
\value{
\code{query} with updated \code{txName}, \code{geneName}
  column slots
}
\description{
Features in \code{query} are annotated with respect to transcript
features in \code{subject}.
}
\details{
Annotation is performed at the gene and transcript level.
For transcript-level annotation, query features are assigned all
transcript names associated with matching subject features.
For gene-level annotation, query features are assigned all gene
names associated with subject features that belong to the same
gene (connected component of the splice graph) as matching
query features.

Feature matching is performed as follows:
Query splice junctions are matched with identical subject splice
junctions. Query splice sites are matched with splice sites
implied by subject splice junctions. Query exon bins are matched with
overlapping subject exons. Spliced boundaries of query exon bins
must match spliced subject exon boundaries. Query exon bins cannot
extend across spliced subject exon boundaries.
}
\examples{
sgf_annotated <- annotate(sgf_pred, txf_ann)
sgv_annotated <- annotate(sgv_pred, txf_ann)
}
\author{
Leonard Goldstein
}

