% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slots-generics.R, R/slots-methods.R
\docType{methods}
\name{assays}
\alias{FPKM}
\alias{FPKM,SGFeatureCounts-method}
\alias{FPKM<-}
\alias{FPKM<-,SGFeatureCounts-method}
\alias{assays}
\alias{counts,SGFeatureCounts-method}
\alias{counts,SGVariantCounts-method}
\alias{counts3p}
\alias{counts3p,SGVariantCounts-method}
\alias{counts3p<-}
\alias{counts3p<-,SGVariantCounts-method}
\alias{counts3pEvent}
\alias{counts3pEvent,SGVariantCounts-method}
\alias{counts3pEvent<-}
\alias{counts3pEvent<-,SGVariantCounts-method}
\alias{counts5p}
\alias{counts5p,SGVariantCounts-method}
\alias{counts5p<-}
\alias{counts5p<-,SGVariantCounts-method}
\alias{counts5pEvent}
\alias{counts5pEvent,SGVariantCounts-method}
\alias{counts5pEvent<-}
\alias{counts5pEvent<-,SGVariantCounts-method}
\alias{counts<-,SGFeatureCounts-method}
\alias{counts<-,SGVariantCounts-method}
\alias{variantFreq}
\alias{variantFreq,SGVariantCounts-method}
\alias{variantFreq<-}
\alias{variantFreq<-,SGVariantCounts-method}
\title{Accessing and replacing assay data}
\usage{
FPKM(object)

FPKM(object) <- value

counts5p(object)

counts5p(object) <- value

counts3p(object)

counts3p(object) <- value

counts5pEvent(object)

counts5pEvent(object) <- value

counts3pEvent(object)

counts3pEvent(object) <- value

variantFreq(object)

variantFreq(object) <- value

\S4method{counts}{SGFeatureCounts}(object)

\S4method{counts}{SGFeatureCounts}(object) <- value

\S4method{FPKM}{SGFeatureCounts}(object)

\S4method{FPKM}{SGFeatureCounts}(object) <- value

\S4method{counts5p}{SGVariantCounts}(object)

\S4method{counts5p}{SGVariantCounts}(object) <- value

\S4method{counts3p}{SGVariantCounts}(object)

\S4method{counts3p}{SGVariantCounts}(object) <- value

\S4method{counts5pEvent}{SGVariantCounts}(object)

\S4method{counts5pEvent}{SGVariantCounts}(object) <- value

\S4method{counts3pEvent}{SGVariantCounts}(object)

\S4method{counts3pEvent}{SGVariantCounts}(object) <- value

\S4method{variantFreq}{SGVariantCounts}(object)

\S4method{variantFreq}{SGVariantCounts}(object) <- value

\S4method{counts}{SGVariantCounts}(object)

\S4method{counts}{SGVariantCounts}(object) <- value
}
\arguments{
\item{object}{Object containing assay data}

\item{value}{Replacement value}
}
\value{
Assay data for accessor functions, updated object for replacement
functions.
}
\description{
Accessor and replacement functions for assay data.
}
\details{
\code{Counts} objects defined in the \code{SGSeq} package contain
different types of assay data. For example, class \code{SGFeatureCounts}
contains assays \code{counts} and \code{FPKM}.

To facilitate accessing and modifying assays, for each assay
there exists a function with name identical to the assay name
that can be used to access and modify it (see examples).
}
\examples{
x <- counts(sgfc_pred)
y <- FPKM(sgfc_pred)
}
\author{
Leonard Goldstein
}

