% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slots-generics.R, R/slots-methods.R
\docType{methods}
\name{assays}
\alias{FPKM}
\alias{FPKM,SGFeatureCounts-method}
\alias{FPKM,SGVariantCounts-method}
\alias{FPKM<-}
\alias{FPKM<-,SGFeatureCounts-method}
\alias{assays}
\alias{counts,SGFeatureCounts-method}
\alias{counts,SGVariantCounts-method}
\alias{counts<-,SGFeatureCounts-method}
\alias{counts<-,SGVariantCounts-method}
\alias{variantFreq}
\alias{variantFreq,SGVariantCounts-method}
\alias{variantFreq<-}
\alias{variantFreq<-,SGVariantCounts-method}
\title{Accessing and replacing assay data}
\usage{
FPKM(object, ...)

FPKM(object, ...) <- value

variantFreq(object)

variantFreq(object) <- value

\S4method{counts}{SGFeatureCounts}(object)

\S4method{counts}{SGFeatureCounts}(object) <- value

\S4method{FPKM}{SGFeatureCounts}(object)

\S4method{FPKM}{SGFeatureCounts}(object) <- value

\S4method{counts}{SGVariantCounts}(object, option)

\S4method{counts}{SGVariantCounts}(object, option) <- value

\S4method{FPKM}{SGVariantCounts}(object, option, min_anchor = 1)

\S4method{variantFreq}{SGVariantCounts}(object)

\S4method{variantFreq}{SGVariantCounts}(object) <- value
}
\arguments{
\item{object}{Object containing assay data}

\item{...}{Arguments passed to method for \code{SGVariantCounts} objects}

\item{value}{Replacement value}

\item{option}{For \code{SGVariantCounts} objects, \code{option} specifies
whether output should be based on counts of reads compatible with the
variant at the start (\dQuote{variant5p}), end (\dQuote{variant3p}) or
either (\dQuote{variant}), or whether output should be based on counts of
reads compatible with any variant belonging to the same event
(\dQuote{event5p} or \dQuote{event3p})}

\item{min_anchor}{For \code{SGVariantCounts} objects, \code{min_anchor}
specifies the minimum anchor length when computing FPKM values}
}
\value{
Assay data for accessor functions or updated object for
replacement functions.
}
\description{
Accessor and replacement functions for assay data.
}
\details{
Functions \code{counts} and \code{FPKM} are used to extract counts and
FPKM values from \code{SGFeatureCounts} and \code{SGVariantCounts}
objects. Function \code{variantFreq} is used to access relative usage
estimates from \code{SGVariantCounts} objects.
}
\examples{
x <- counts(sgfc_pred)
y <- FPKM(sgfc_pred)
u <- counts(sgvc_pred, "variant5p")
v <- FPKM(sgvc_pred, "variant5p")
}
\author{
Leonard Goldstein
}

