% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slots-generics.R, R/slots-methods.R
\docType{methods}
\name{assays}
\alias{FPKM}
\alias{FPKM,SGFeatureCounts-method}
\alias{FPKM<-}
\alias{FPKM<-,SGFeatureCounts-method}
\alias{assays}
\alias{counts,SGFeatureCounts-method}
\alias{counts<-,SGFeatureCounts-method}
\alias{countsTotal}
\alias{countsTotal,SGVariantCounts-method}
\alias{countsTotal3p}
\alias{countsTotal3p,SGVariantCounts-method}
\alias{countsTotal3p<-}
\alias{countsTotal3p<-,SGVariantCounts-method}
\alias{countsTotal5p}
\alias{countsTotal5p,SGVariantCounts-method}
\alias{countsTotal5p<-}
\alias{countsTotal5p<-,SGVariantCounts-method}
\alias{countsTotal<-}
\alias{countsTotal<-,SGVariantCounts-method}
\alias{countsVariant}
\alias{countsVariant,SGVariantCounts-method}
\alias{countsVariant3p}
\alias{countsVariant3p,SGVariantCounts-method}
\alias{countsVariant3p<-}
\alias{countsVariant3p<-,SGVariantCounts-method}
\alias{countsVariant5p}
\alias{countsVariant5p,SGVariantCounts-method}
\alias{countsVariant5p<-}
\alias{countsVariant5p<-,SGVariantCounts-method}
\alias{countsVariant<-}
\alias{countsVariant<-,SGVariantCounts-method}
\alias{variantFreq}
\alias{variantFreq,SGVariantCounts-method}
\alias{variantFreq<-}
\alias{variantFreq<-,SGVariantCounts-method}
\title{Accessing and replacing assay data}
\usage{
FPKM(object)

FPKM(object) <- value

countsVariant5p(object)

countsVariant5p(object) <- value

countsVariant3p(object)

countsVariant3p(object) <- value

countsTotal5p(object)

countsTotal5p(object) <- value

countsTotal3p(object)

countsTotal3p(object) <- value

countsVariant(object)

countsVariant(object) <- value

countsTotal(object)

countsTotal(object) <- value

variantFreq(object)

variantFreq(object) <- value

\S4method{counts}{SGFeatureCounts}(object)

\S4method{counts}{SGFeatureCounts}(object) <- value

\S4method{FPKM}{SGFeatureCounts}(object)

\S4method{FPKM}{SGFeatureCounts}(object) <- value

\S4method{countsVariant5p}{SGVariantCounts}(object)

\S4method{countsVariant5p}{SGVariantCounts}(object) <- value

\S4method{countsVariant3p}{SGVariantCounts}(object)

\S4method{countsVariant3p}{SGVariantCounts}(object) <- value

\S4method{countsTotal5p}{SGVariantCounts}(object)

\S4method{countsTotal5p}{SGVariantCounts}(object) <- value

\S4method{countsTotal3p}{SGVariantCounts}(object)

\S4method{countsTotal3p}{SGVariantCounts}(object) <- value

\S4method{variantFreq}{SGVariantCounts}(object)

\S4method{variantFreq}{SGVariantCounts}(object) <- value

\S4method{countsVariant}{SGVariantCounts}(object)

\S4method{countsVariant}{SGVariantCounts}(object) <- value

\S4method{countsTotal}{SGVariantCounts}(object)

\S4method{countsTotal}{SGVariantCounts}(object) <- value
}
\arguments{
\item{object}{Object containing assay data}

\item{value}{Replacement value}
}
\value{
Assay data for accessor functions, updated object for replacement
functions.
}
\description{
Accessor and replacement functions for assay data.
}
\details{
\code{Counts} objects defined in the \code{SGSeq} package contain
different types of assay data. For example, class \code{SGFeatureCounts}
contains assays \code{counts} and \code{FPKM}.

To facilitate accessing and modifying assays, for each assay
there exists a function with name identical to the assay name
that can be used to access and modify it (see examples).
}
\examples{
x <- counts(sgfc)
y <- FPKM(sgfc)
}
\author{
Leonard Goldstein
}

