% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{exportFeatures}
\alias{exportFeatures}
\title{Export to BED format}
\usage{
exportFeatures(features, file)
}
\arguments{
\item{features}{\code{TxFeatures} or \code{SGFeatures} object}

\item{file}{Character string specifying output file}
}
\value{
\code{NULL}
}
\description{
Export features to BED format. Splice sites are not included.
}
\examples{
\dontrun{
exportFeatures(txf, "txf.bed")
exportFeatures(sgf, "sgf.bed")
}
}
\author{
Leonard Goldstein
}

