% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBamInfo}
\alias{getBamInfo}
\title{Obtain alignment information from BAM files}
\usage{
getBamInfo(sample_info, yieldSize = NULL, BPPARAM = MulticoreParam(1))
}
\arguments{
\item{sample_info}{\code{data.frame} with sample information including
mandatory character columns \dQuote{sample_name} and \dQuote{file_bam}.}

\item{yieldSize}{Number of records used for obtaining alignment
information, or \code{NULL} for all records}

\item{BPPARAM}{\code{BiocParallelParam} for processing samples in
parallel, defaults to \code{MulticoreParam(1)}}
}
\value{
\code{sample_info} with additional columns \dQuote{paired_end},
  \dQuote{read_length}, \dQuote{frag_length}, and \dQuote{lib_size}
  if \code{yieldSize} is \code{NULL}
}
\description{
Obtain paired-end status, median aligned read length,
median aligned insert size and library size from BAM file.
}
\details{
Alignment information can be inferred from a subset of BAM records
by setting the number of records via argument \code{yieldSize}.
Note that library size can only be obtained if \code{yieldSize} is {NULL}.
}
\examples{
dir <- system.file("extdata", package = "SGSeq")
si$file_bam <- file.path(dir, "bams", si$file_bam)
si <- si[, c("sample_name", "file_bam")]
si_complete <- getBamInfo(si)
}
\author{
Leonard Goldstein
}

