% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSGFeatureCounts}
\alias{getSGFeatureCounts}
\title{Compatible counts for splice graph features from BAM files}
\usage{
getSGFeatureCounts(sample_info, features, cores_per_sample = 1,
  BPPARAM = MulticoreParam(1))
}
\arguments{
\item{sample_info}{\code{data.frame} with sample information.
Required columns are \dQuote{sample_name}, \dQuote{file_bam},
\dQuote{paired_end}, \dQuote{read_length}, \dQuote{frag_length}
and \dQuote{lib_size}. Alignment information can be obtained with
function \code{getBamInfo}.}

\item{features}{\code{SGFeatures} object}

\item{cores_per_sample}{Number of cores per sample}

\item{BPPARAM}{\code{BiocParallelParam} for processing samples in parallel,
defaults to \code{MulticoreParam(1)}}
}
\value{
An \code{SGFeatureCounts} object
}
\description{
Compatible counts are obtained for each sample and combined into
an \code{SGFeatureCounts} object.
}
\examples{
dir <- system.file("extdata", package = "SGSeq")
si$file_bam <- file.path(dir, "bams", si$file_bam)
sgfc <- getSGFeatureCounts(si, sgf)
}
\author{
Leonard Goldstein
}

