% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSGFeatureCountsPerSample}
\alias{getSGFeatureCountsPerSample}
\title{Compatible fragment counts for splice graph features}
\usage{
getSGFeatureCountsPerSample(features, file_bam, paired_end,
  retain_coverage = FALSE, cores = 1)
}
\arguments{
\item{features}{\code{SGFeatures} object}

\item{file_bam}{BAM file with genomic RNA-seq read alignments}

\item{paired_end}{Logical, \code{TRUE} for paired-end data,
\code{FALSE} for single-end data}

\item{retain_coverage}{Logical indicating whether coverages for each
exon should be retained as an \code{IntegerList} in elementMetadata
column \dQuote{coverage}. This allows filtering of features
using more stringent criteria after the initial prediction.}

\item{cores}{Number of cores available for parallel processing}
}
\value{
Numeric vector of compatible fragment counts
}
\description{
Obtain counts of compatible fragments for splice graph features.
}
\author{
Leonard Goldstein
}
\keyword{internal}

