% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{getSGVariantCounts}
\alias{getSGVariantCounts}
\title{Representative counts and frequency estimates for
  transcript variants}
\usage{
getSGVariantCounts(variants, object = NULL, features = rowRanges(object),
  cores = 1, sample_info = NULL, verbose = FALSE, cores_per_sample = 1,
  BPPARAM = SerialParam())
}
\arguments{
\item{variants}{\code{SGVariants} object}

\item{object}{\code{SGFeatureCounts} object}

\item{features}{\code{SGFeatures} object that must include all features
included in featureID5p(variants) and featureID3p(variants)}

\item{cores}{Number of cores available for parallel processing
when obtaining counts from an \code{SGFeatureCounts} object}

\item{sample_info}{Data frame with sample information.
Required columns are \dQuote{sample_name}, \dQuote{file_bam},
\dQuote{paired_end}, \dQuote{read_length}, \dQuote{frag_length}
and \dQuote{lib_size}. Alignment information can be obtained with
function \code{getBamInfo}.}

\item{verbose}{If \code{TRUE}, generate messages indicating progress}

\item{cores_per_sample}{Number of cores per sample}

\item{BPPARAM}{\code{BiocParallelParam} for processing samples in parallel,
defaults to \code{SerialParam()}}
}
\value{
An \code{SGVariantCounts} object
}
\description{
For transcript variants, obtain counts of compatible fragments
extending across the start and/or end of each variant.
Variant frequencies are estimated based on representive counts.
}
\examples{
sgvc_from_sgfc <- getSGVariantCounts(sgv, sgfc)
path <- system.file("extdata", package = "SGSeq")
si$file_bam <- file.path(path, "bams", si$file_bam)
sgvc_from_bam <- getSGVariantCounts(sgv, features = sgf, sample_info = si)
}
\author{
Leonard Goldstein
}

