% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{getSGVariantCounts}
\alias{getSGVariantCounts}
\title{Representative counts and frequency estimates for
  splice variants}
\usage{
getSGVariantCounts(variants, object = NULL, features = NULL,
  sample_info = NULL, verbose = FALSE, cores = 1)
}
\arguments{
\item{variants}{\code{SGVariants} object}

\item{object}{\code{SGFeatureCounts} object}

\item{features}{\code{SGFeatures} object that must include all features
included in featureID5p(variants) and featureID3p(variants)}

\item{sample_info}{Data frame with sample information.
Required columns are \dQuote{sample_name}, \dQuote{file_bam},
\dQuote{paired_end}, \dQuote{read_length}, \dQuote{frag_length}
and \dQuote{lib_size}. Library information can be obtained with
function \code{getBamInfo}.}

\item{verbose}{If \code{TRUE}, generate messages indicating progress}

\item{cores}{Number of cores available for parallel processing}
}
\value{
An \code{SGVariantCounts} object
}
\description{
For splice variants obtain counts of compatible fragments
extending across the start or end of each variant.
Counts can be obtained from an \code{SGFeatureCounts} object
or from BAM files. Only one of the two arguments \code{object}
and \code{sample_info} must be specified. Splice variant
frequencies are estimated based on representive counts.
}
\examples{
sgvc_from_sgfc <- getSGVariantCounts(sgv, sgfc)
path <- system.file("extdata", package = "SGSeq")
si$file_bam <- file.path(path, "bams", si$file_bam)
sgvc_from_bam <- getSGVariantCounts(sgv, features = sgf, sample_info = si)
}
\author{
Leonard Goldstein
}

