% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{makeSGFeatureCounts}
\alias{makeSGFeatureCounts}
\title{Create \code{SGFeatureCounts} object}
\usage{
makeSGFeatureCounts(rowRanges, colData, counts, min_anchor = 1)
}
\arguments{
\item{rowRanges}{\code{SGFeatures} object}

\item{colData}{Data frame with sample information}

\item{counts}{Integer matrix of counts}

\item{min_anchor}{Integer specifiying minimum anchor length}
}
\value{
\code{SGFeatureCounts} object
}
\description{
Create \code{SGFeatureCounts} object from rowRanges, colData and counts.
}
\examples{
sgfc <- makeSGFeatureCounts(sgf_pred, si,
  matrix(0L, length(sgf_pred), nrow(si)))
}
\author{
Leonard Goldstein
}

