% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/counts.R
\name{makeSGFeatureCounts}
\alias{makeSGFeatureCounts}
\title{Create \code{SGFeatureCounts} object}
\usage{
makeSGFeatureCounts(rowRanges, colData, counts)
}
\arguments{
\item{rowRanges}{An \code{SGFeatures} object}

\item{colData}{Data frame with sample information}

\item{counts}{Integer matrix of counts}
}
\value{
An \code{SGFeatureCounts} object
}
\description{
Create \code{SGFeatureCounts} object from rowRanges, colData and counts.
}
\examples{
sgfc <- makeSGFeatureCounts(sgf, si, matrix(0L, length(sgf), nrow(si)))
}
\author{
Leonard Goldstein
}

