% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeVariantNames}
\alias{makeVariantNames}
\title{Create interpretable transcript variant names}
\usage{
makeVariantNames(variants)
}
\arguments{
\item{variants}{\code{TxVariants} object}
}
\value{
Character vector with transcript variant names
}
\description{
This function creates interpretable transcript variant names
taking the format GENE_EVENT_VARIANT/ORDER_TYPE.
GENE is based on geneName if available, and geneID otherwise.
EVENT and TYPE are based on eventID and variantType, respectively.
VARIANT enumerates transcript variants in the same event, while
ORDER indicates the total number of variants in the event (e.g. 1/2
refers to the first out of two transcript variants in the event).
}
\examples{
makeVariantNames(txv)
}
\author{
Leonard Goldstein
}

