% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotVariants}
\alias{plotVariants}
\title{Plot splice graph and heatmap of splice variant frequencies}
\usage{
plotVariants(x, eventID = NULL, toscale = c("exon", "none", "gene"),
  color = "gray", color_novel = color, color_alpha = 0.8,
  color_labels = FALSE, border = "fill", cexLab = 1, cexExon = 1,
  ypos = 0.5, score = NULL, score_color = "darkblue", score_ylim = NULL,
  score_ypos = c(0.3, 0.1), score_nbin = 400, ranges = NULL,
  ranges_color = "darkblue", ranges_ypos = c(0.1, 0.1), main = NULL,
  cexMain = 1, tx_view = FALSE, tx_dist = 0.1, tx_cex = 1,
  transform = function(x) {     x }, Rowv = NULL, distfun = dist,
  hclustfun = hclust, margin = 0.2, RowSideColors = NULL,
  square = FALSE, cexRow = 1, cexCol = 1, labRow = colnames(x),
  col = colorRampPalette(c("black", "gold"))(256), zlim = c(0, 1),
  heightTopPanel = 0.3, expand_variants = FALSE)
}
\arguments{
\item{x}{\code{SGVariantCounts} object}

\item{eventID}{Single event identifier used to subset \code{x}}

\item{toscale}{Controls which parts of the splice graph are drawn to
scale. Possible values are \dQuote{none} (exonic and intronic regions
have constant length), \dQuote{exon} (exonic regions are drawn to scale)
and \dQuote{gene} (both exonic and intronic regions are drawn to scale).}

\item{color}{Color used for plotting the splice graph. Ignored if features
elementMetadata column \dQuote{color} is not \code{NULL}.}

\item{color_novel}{Features with missing annotation are
highlighted in \code{color_novel}. Ignored if features
elementMetadata column \dQuote{color} is not \code{NULL}.}

\item{color_alpha}{Controls color transparency}

\item{color_labels}{Logical indicating whether label colors should
be the same as feature colors}

\item{border}{Determines the color of exon borders, can be \dQuote{fill}
(same as exon color), \dQuote{none} (no border) or a valid color name}

\item{cexLab}{Scale factor for feature labels}

\item{cexExon}{Scale factor for exon height}

\item{ypos}{Numeric value indicating vertical position of splice graph
(exon bins) specified as fraction of height of the plotting region
(not supported for \code{tx_view = TRUE})}

\item{score}{\code{RLeList} containing nucleotide-level scores
to be plotted with the splice graph}

\item{score_color}{Color used for plotting scores}

\item{score_ylim}{y-axis range used for plotting scores}

\item{score_ypos}{Numeric vector of length two, indicating the vertical
position and height of the score panel, specificed as fraction of the
height of the plotting region}

\item{score_nbin}{Number of bins for plotting scores}

\item{ranges}{\code{GRangesList} to be plotted with the splice graph}

\item{ranges_color}{Color used for plotting ranges}

\item{ranges_ypos}{Numeric vector of length two, indicating the vertical
position and height of the ranges panel, specificed as fraction of the
height of the plotting region}

\item{main}{Plot title}

\item{cexMain}{Scale factor for plot title}

\item{tx_view}{Plot transcripts instead of splice graph (experimental)}

\item{tx_dist}{Vertical distance between transcripts as fraction of height
of plotting region}

\item{tx_cex}{Scale factor for transcript labels}

\item{transform}{Transformation applied to splice variant frequencies}

\item{Rowv}{Determines order of rows. Either a vector of values used to
reorder rows, or \code{NA} to suppress reordering, or \code{NULL} for
hierarchical clustering.}

\item{distfun}{Distance function used for hierarchical clustering
of rows (samples)}

\item{hclustfun}{Clustering function used for hierarchical clustering
of rows (samples)}

\item{margin}{Width of right-hand margin as fraction of width of the
graphics device. Ignored if \code{square} is \code{TRUE}.}

\item{RowSideColors}{Character vector (or list of character vectors)
with length(s) equal to \code{ncol(x)} containing color names for
horizontal side bars for sample annotation}

\item{square}{Logical, if \code{TRUE} margins are set such that
cells in the heatmap are square}

\item{cexRow}{Scale factor for row (sample) labels}

\item{cexCol}{Scale factor for column (feature) labels}

\item{labRow}{Character vector of row (sample) labels}

\item{col}{Heatmap colors}

\item{zlim}{Range of values for which colors should be plotted,
if \code{NULL} range of finite values}

\item{heightTopPanel}{Height of top panel as fraction of height of the
graphics device}

\item{expand_variants}{Experimental option - leave set to \code{FALSE}}
}
\value{
\code{data.frame} with information on exon bins and
  splice junctions included in the splice graph
}
\description{
Plot splice graph and heatmap of splice variant frequencies
}
\examples{
\dontrun{
sgvc_annotated <- annotate(sgvc, txf)
plotVariants(sgvc_annotated)
}
}
\author{
Leonard Goldstein
}

