% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictCandidatesInternal}
\alias{predictCandidatesInternal}
\title{Identify candidate internal exons}
\usage{
predictCandidatesInternal(junctions, islands)
}
\arguments{
\item{junctions}{\code{IRanges} of splice junctions}

\item{islands}{\code{IRanges} of genomic regions with minimal read
coverage required for internal exon prediction}
}
\value{
\code{IRanges} of candidate internal exons
}
\description{
Identify candidate internal exons based on previously identified
splice junctions and regions with minimal read coverage required
for internal exon prediction.
}
\author{
Leonard Goldstein
}
\keyword{internal}

