% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictCandidatesTerminal}
\alias{predictCandidatesTerminal}
\title{Identify candidate terminal exons}
\usage{
predictCandidatesTerminal(junctions, islands, type = c("exon_L", "exon_R"))
}
\arguments{
\item{junctions}{\code{IRanges} of splice junctions}

\item{islands}{\code{IRanges} of genomic regions with minimal read
coverage required for internal exon prediction}

\item{type}{Character string indicating whether terminal exons
should be identified to the left (\dQuote{exon_L}) or right
(\dQuote{exon_R}) of provided splice junctions}
}
\value{
\code{IRanges} of candidate terminal exons
}
\description{
Identify candidate terminal exons based on previously identified
splice junctions and regions with minimal read coverage required
for terminal exon prediction.
}
\author{
Leonard Goldstein
}
\keyword{internal}

