% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-prediction.R
\name{predictExonsInternal}
\alias{predictExonsInternal}
\title{Identify internal exons}
\usage{
predictExonsInternal(candidates, frag_exonic, frag_intron, relCov,
  include_counts, retain_coverage)
}
\arguments{
\item{candidates}{\code{IRanges} of candidate internal exons}

\item{frag_exonic}{\code{IRangesList} with exonic regions from alignments}

\item{frag_intron}{\code{IRangesList} with introns implied by spliced
alignments}

\item{relCov}{Minimum relative coverage required for exon prediction}

\item{include_counts}{Logical indicating whether counts of
compatible fragments should be included in elementMetadata column
\dQuote{N}}

\item{retain_coverage}{Logical indicating whether coverages for each
exon should be retained as an \code{IntegerList} in elementMetadata
column \dQuote{coverage}. This allows filtering of features
using more stringent criteria after the initial prediction.}
}
\value{
\code{IRanges} of internal exons with elementMetadata column
  \dQuote{type} and optionally \dQuote{N} for
  \code{include_counts = TRUE}, \dQuote{N_splicesite},
  \dQuote{coverage} for \code{retain_coverage = TRUE}
}
\description{
Identify internal exons based on candidate internal exons and
compatible read coverage.
}
\author{
Leonard Goldstein
}
\keyword{internal}

