% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictSpliced}
\alias{predictSpliced}
\title{\code{IRanges}-based identification of splice junctions and exons}
\usage{
predictSpliced(frag_exonic, frag_intron, min_junction_count, psi, beta, gamma,
  include_counts, retain_coverage, junctions_only)
}
\arguments{
\item{frag_exonic}{\code{IRangesList} with exonic regions from alignments}

\item{frag_intron}{\code{IRangesList} with introns implied by spliced
alignments}

\item{min_junction_count}{Minimum fragment count required for a splice
junction to be included. If specified, argument \code{alpha} is ignored.}

\item{psi}{Minimum splice frequency required for a splice junction
to be included}

\item{beta}{Minimum relative coverage required for an internal exon
to be included}

\item{gamma}{Minimum relative coverage required for a terminal exon
to be included}

\item{include_counts}{Logical indicating whether counts of
compatible fragments should be included in elementMetadata column
\dQuote{N}}

\item{retain_coverage}{Logical indicating whether coverages for each
exon should be retained as an \code{IntegerList} in elementMetadata
column \dQuote{coverage}. This allows filtering of features
using more stringent criteria after the initial prediction.}

\item{junctions_only}{Logical indicating whether predictions
should be limited to identification of splice junctions only}
}
\value{
\code{IRanges} with predicted features
}
\description{
\code{IRanges}-based identification of splice junctions and exons
}
\author{
Leonard Goldstein
}
\keyword{internal}

