% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-filter.R
\name{removeExonsIsolated}
\alias{removeExonsIsolated}
\title{Remove exons with no flanking splice junctions}
\usage{
removeExonsIsolated(features)
}
\arguments{
\item{features}{\code{TxFeatures} object}
}
\value{
\code{TxFeatures} object with filtered features
}
\description{
Remove exons with no flanking splice junctions
}
\author{
Leonard Goldstein
}
\keyword{internal}

