% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeExonsIsolated}
\alias{removeExonsIsolated}
\title{Remove exons with no flanking splice junctions}
\usage{
removeExonsIsolated(features, return_index = FALSE)
}
\arguments{
\item{features}{\code{TxFeatures} object}

\item{return_index}{Logical indicating whether indices of retained
features should be returned instead of filtered features}
}
\value{
\code{TxFeatures} object with filtered features,
  or indices of retained features if \code{return_index = TRUE}
}
\description{
Remove exons with no flanking splice junctions
}
\author{
Leonard Goldstein
}
\keyword{internal}

