% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{sgfc}
\alias{sgfc}
\title{Example splice graph feature counts}
\format{An \code{SGFeatureCounts} object.}
\description{
Compatible counts and FPKMs for predicted FBXO31 splice graph features.
}
\author{
Leonard Goldstein
}
\keyword{internal}

