% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SGFeatureCounts}
\alias{SGFeatureCounts}
\title{Splice graph feature counts}
\usage{
SGFeatureCounts(x)
}
\arguments{
\item{x}{\code{RangedSummarizedExperiment} with \code{SGFeatures}
as \code{rowRanges} and assays \dQuote{counts} and \dQuote{FPKM}}
}
\value{
\code{SGFeatureCounts} object
}
\description{
Creates an instance of S4 class \code{SGFeatureCounts} for storing
compatible splice graph feature counts.
}
\examples{
sgfc <- SGFeatureCounts()
}
\author{
Leonard Goldstein
}
