% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SGFeatureCounts}
\alias{SGFeatureCounts}
\title{Constructor function for S4 class \code{SGFeatureCounts}}
\usage{
SGFeatureCounts(x)
}
\arguments{
\item{x}{\code{RangedSummarizedExperiment} with \code{SGFeatures}
as \code{rowRanges} and assays \dQuote{counts}, \dQuote{FPKM}}
}
\value{
An \code{SGFeatureCounts} object
}
\description{
Creates an instance of S4 class \code{SGFeatureCounts} for storing
compatible splice graph feature counts.
}
\examples{
sgfc <- SGFeatureCounts()
}
\author{
Leonard Goldstein
}

