% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{sgv}
\alias{sgv}
\title{Example splice variants}
\format{A \code{SGVariants} object.}
\description{
Splice variants for FBXO31, based on splice graph features predicted
from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

