% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SGVariantCounts}
\alias{SGVariantCounts}
\title{Constructor function for S4 class \code{SGFeatureCounts}}
\usage{
SGVariantCounts(x)
}
\arguments{
\item{x}{\code{RangedSummarizedExperiment} with \code{SGVariants}
as \code{rowRanges} and appropriate assays}
}
\value{
A \code{SGVariantCounts} object
}
\description{
Creates an instance of S4 class \code{SGVariantCounts} for storing
representative splice variant counts.
}
\examples{
sgvc <- SGVariantCounts()
}
\author{
Leonard Goldstein
}

