% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{sgvc}
\alias{sgvc}
\title{Example splice variant counts}
\format{A \code{SGVariantCounts} object.}
\description{
Splice variant counts and frequencies for FBXO31. Splice graph
features were predicted and counts obtained from example BAM files.
}
\author{
Leonard Goldstein
}
\keyword{internal}

