% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{slots}
\alias{closed3p}
\alias{closed3p,TxVariantCounts-method}
\alias{closed3p,TxVariants-method}
\alias{closed3p<-}
\alias{closed3p<-,TxVariantCounts-method}
\alias{closed3p<-,TxVariants-method}
\alias{closed5p}
\alias{closed5p,TxVariantCounts-method}
\alias{closed5p,TxVariants-method}
\alias{closed5p<-}
\alias{closed5p<-,TxVariantCounts-method}
\alias{closed5p<-,TxVariants-method}
\alias{eventID}
\alias{eventID,TxVariantCounts-method}
\alias{eventID,TxVariants-method}
\alias{eventID<-}
\alias{eventID<-,TxVariantCounts-method}
\alias{eventID<-,TxVariants-method}
\alias{featureID}
\alias{featureID,Counts-method}
\alias{featureID,Paths-method}
\alias{featureID,SGFeatures-method}
\alias{featureID3p}
\alias{featureID3p,TxVariantCounts-method}
\alias{featureID3p,TxVariants-method}
\alias{featureID3p<-}
\alias{featureID3p<-,TxVariantCounts-method}
\alias{featureID3p<-,TxVariants-method}
\alias{featureID5p}
\alias{featureID5p,TxVariantCounts-method}
\alias{featureID5p,TxVariants-method}
\alias{featureID5p<-}
\alias{featureID5p<-,TxVariantCounts-method}
\alias{featureID5p<-,TxVariants-method}
\alias{featureID<-}
\alias{featureID<-,Counts-method}
\alias{featureID<-,Paths-method}
\alias{featureID<-,SGFeatures-method}
\alias{from}
\alias{from,Paths-method}
\alias{from,TxVariantCounts-method}
\alias{from<-}
\alias{from<-,Paths-method}
\alias{from<-,TxVariantCounts-method}
\alias{geneID}
\alias{geneID,Counts-method}
\alias{geneID,Paths-method}
\alias{geneID,SGFeatures-method}
\alias{geneID<-}
\alias{geneID<-,Counts-method}
\alias{geneID<-,Paths-method}
\alias{geneID<-,SGFeatures-method}
\alias{geneName}
\alias{geneName,Counts-method}
\alias{geneName,Features-method}
\alias{geneName,Paths-method}
\alias{geneName<-}
\alias{geneName<-,Counts-method}
\alias{geneName<-,Features-method}
\alias{geneName<-,Paths-method}
\alias{segmentID}
\alias{segmentID,Paths-method}
\alias{segmentID,TxVariantCounts-method}
\alias{segmentID<-}
\alias{segmentID<-,Paths-method}
\alias{segmentID<-,TxVariantCounts-method}
\alias{slots}
\alias{splice3p}
\alias{splice3p,SGFeatureCounts-method}
\alias{splice3p,SGFeatures-method}
\alias{splice3p,TxSegments-method}
\alias{splice3p<-}
\alias{splice3p<-,SGFeatureCounts-method}
\alias{splice3p<-,SGFeatures-method}
\alias{splice3p<-,TxSegments-method}
\alias{splice5p}
\alias{splice5p,SGFeatureCounts-method}
\alias{splice5p,SGFeatures-method}
\alias{splice5p,TxSegments-method}
\alias{splice5p<-}
\alias{splice5p<-,SGFeatureCounts-method}
\alias{splice5p<-,SGFeatures-method}
\alias{splice5p<-,TxSegments-method}
\alias{to}
\alias{to,Paths-method}
\alias{to,TxVariantCounts-method}
\alias{to<-}
\alias{to<-,Paths-method}
\alias{to<-,TxVariantCounts-method}
\alias{txName}
\alias{txName,Counts-method}
\alias{txName,Features-method}
\alias{txName,Paths-method}
\alias{txName<-}
\alias{txName<-,Counts-method}
\alias{txName<-,Features-method}
\alias{txName<-,Paths-method}
\alias{type}
\alias{type,Counts-method}
\alias{type,Features-method}
\alias{type,Paths-method}
\alias{type<-}
\alias{type<-,Counts-method}
\alias{type<-,Features-method}
\alias{type<-,Paths-method}
\alias{variantID}
\alias{variantID,TxVariantCounts-method}
\alias{variantID,TxVariants-method}
\alias{variantID<-}
\alias{variantID<-,TxVariantCounts-method}
\alias{variantID<-,TxVariants-method}
\alias{variantName}
\alias{variantName,TxVariantCounts-method}
\alias{variantName,TxVariants-method}
\alias{variantName<-}
\alias{variantName<-,TxVariantCounts-method}
\alias{variantName<-,TxVariants-method}
\alias{variantType}
\alias{variantType,TxVariantCounts-method}
\alias{variantType,TxVariants-method}
\alias{variantType<-}
\alias{variantType<-,TxVariantCounts-method}
\alias{variantType<-,TxVariants-method}
\title{Accessing and replacing column slots}
\usage{
type(object)

type(object) <- value

txName(object)

txName(object) <- value

geneName(object)

geneName(object) <- value

featureID(object)

featureID(object) <- value

geneID(object)

geneID(object) <- value

splice5p(object)

splice5p(object) <- value

splice3p(object)

splice3p(object) <- value

from(object)

from(object) <- value

to(object)

to(object) <- value

segmentID(object)

segmentID(object) <- value

variantID(object)

variantID(object) <- value

eventID(object)

eventID(object) <- value

closed5p(object)

closed5p(object) <- value

closed3p(object)

closed3p(object) <- value

variantType(object)

variantType(object) <- value

variantName(object)

variantName(object) <- value

featureID5p(object)

featureID5p(object) <- value

featureID3p(object)

featureID3p(object) <- value

\S4method{type}{Features}(object)

\S4method{type}{Paths}(object)

\S4method{type}{Counts}(object)

\S4method{type}{Features}(object) <- value

\S4method{type}{Paths}(object) <- value

\S4method{type}{Counts}(object) <- value

\S4method{txName}{Features}(object)

\S4method{txName}{Paths}(object)

\S4method{txName}{Counts}(object)

\S4method{txName}{Features}(object) <- value

\S4method{txName}{Paths}(object) <- value

\S4method{txName}{Counts}(object) <- value

\S4method{geneName}{Features}(object)

\S4method{geneName}{Paths}(object)

\S4method{geneName}{Counts}(object)

\S4method{geneName}{Features}(object) <- value

\S4method{geneName}{Paths}(object) <- value

\S4method{geneName}{Counts}(object) <- value

\S4method{featureID}{SGFeatures}(object)

\S4method{featureID}{Paths}(object)

\S4method{featureID}{Counts}(object)

\S4method{featureID}{SGFeatures}(object) <- value

\S4method{featureID}{Paths}(object) <- value

\S4method{featureID}{Counts}(object) <- value

\S4method{geneID}{SGFeatures}(object)

\S4method{geneID}{Paths}(object)

\S4method{geneID}{Counts}(object)

\S4method{geneID}{SGFeatures}(object) <- value

\S4method{geneID}{Paths}(object) <- value

\S4method{geneID}{Counts}(object) <- value

\S4method{splice5p}{SGFeatures}(object)

\S4method{splice5p}{TxSegments}(object)

\S4method{splice5p}{SGFeatureCounts}(object)

\S4method{splice5p}{SGFeatures}(object) <- value

\S4method{splice5p}{TxSegments}(object) <- value

\S4method{splice5p}{SGFeatureCounts}(object) <- value

\S4method{splice3p}{SGFeatures}(object)

\S4method{splice3p}{TxSegments}(object)

\S4method{splice3p}{SGFeatureCounts}(object)

\S4method{splice3p}{SGFeatures}(object) <- value

\S4method{splice3p}{TxSegments}(object) <- value

\S4method{splice3p}{SGFeatureCounts}(object) <- value

\S4method{segmentID}{Paths}(object)

\S4method{segmentID}{TxVariantCounts}(object)

\S4method{segmentID}{Paths}(object) <- value

\S4method{segmentID}{TxVariantCounts}(object) <- value

\S4method{from}{Paths}(object)

\S4method{from}{TxVariantCounts}(object)

\S4method{from}{Paths}(object) <- value

\S4method{from}{TxVariantCounts}(object) <- value

\S4method{to}{Paths}(object)

\S4method{to}{TxVariantCounts}(object)

\S4method{to}{Paths}(object) <- value

\S4method{to}{TxVariantCounts}(object) <- value

\S4method{eventID}{TxVariants}(object)

\S4method{eventID}{TxVariantCounts}(object)

\S4method{eventID}{TxVariants}(object) <- value

\S4method{eventID}{TxVariantCounts}(object) <- value

\S4method{variantID}{TxVariants}(object)

\S4method{variantID}{TxVariantCounts}(object)

\S4method{variantID}{TxVariants}(object) <- value

\S4method{variantID}{TxVariantCounts}(object) <- value

\S4method{closed5p}{TxVariants}(object)

\S4method{closed5p}{TxVariantCounts}(object)

\S4method{closed5p}{TxVariants}(object) <- value

\S4method{closed5p}{TxVariantCounts}(object) <- value

\S4method{closed3p}{TxVariants}(object)

\S4method{closed3p}{TxVariantCounts}(object)

\S4method{closed3p}{TxVariants}(object) <- value

\S4method{closed3p}{TxVariantCounts}(object) <- value

\S4method{variantName}{TxVariants}(object)

\S4method{variantName}{TxVariantCounts}(object)

\S4method{variantName}{TxVariants}(object) <- value

\S4method{variantName}{TxVariantCounts}(object) <- value

\S4method{variantType}{TxVariants}(object)

\S4method{variantType}{TxVariantCounts}(object)

\S4method{variantType}{TxVariants}(object) <- value

\S4method{variantType}{TxVariantCounts}(object) <- value

\S4method{featureID5p}{TxVariants}(object)

\S4method{featureID5p}{TxVariantCounts}(object)

\S4method{featureID5p}{TxVariants}(object) <- value

\S4method{featureID5p}{TxVariantCounts}(object) <- value

\S4method{featureID3p}{TxVariants}(object)

\S4method{featureID3p}{TxVariantCounts}(object)

\S4method{featureID3p}{TxVariants}(object) <- value

\S4method{featureID3p}{TxVariantCounts}(object) <- value
}
\arguments{
\item{object}{Object containing column slot}

\item{value}{Replacement value}
}
\value{
Column value for accessor functions, updated object for
replacement functions.
}
\description{
Accessor and replacement functions for column slots.
}
\details{
S4 classes defined in the \code{SGSeq} package contain columns
with information for each element in the object. For example, class
\code{TxFeatures} contains a column \code{type} that indicates feature
type. The specific columns contained in an object depend on its class.

To facilitate accessing and modifying columns, for each column
there exists a function, with name identical to the column name,
that can be used to access and modify it (see examples).
}
\examples{
head(type(txf))
head(type(sgf))
}
\author{
Leonard Goldstein
}

