% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slots-generics.R, R/slots-methods.R
\docType{methods}
\name{slots}
\alias{closed3p}
\alias{closed3p,SGVariantCounts-method}
\alias{closed3p,SGVariants-method}
\alias{closed3p<-}
\alias{closed3p<-,SGVariantCounts-method}
\alias{closed3p<-,SGVariants-method}
\alias{closed5p}
\alias{closed5p,SGVariantCounts-method}
\alias{closed5p,SGVariants-method}
\alias{closed5p<-}
\alias{closed5p<-,SGVariantCounts-method}
\alias{closed5p<-,SGVariants-method}
\alias{eventID}
\alias{eventID,SGVariantCounts-method}
\alias{eventID,SGVariants-method}
\alias{eventID<-}
\alias{eventID<-,SGVariantCounts-method}
\alias{eventID<-,SGVariants-method}
\alias{featureID}
\alias{featureID,Counts-method}
\alias{featureID,Paths-method}
\alias{featureID,SGFeatures-method}
\alias{featureID3p}
\alias{featureID3p,SGVariantCounts-method}
\alias{featureID3p,SGVariants-method}
\alias{featureID3p<-}
\alias{featureID3p<-,SGVariantCounts-method}
\alias{featureID3p<-,SGVariants-method}
\alias{featureID5p}
\alias{featureID5p,SGVariantCounts-method}
\alias{featureID5p,SGVariants-method}
\alias{featureID5p<-}
\alias{featureID5p<-,SGVariantCounts-method}
\alias{featureID5p<-,SGVariants-method}
\alias{featureID<-}
\alias{featureID<-,Counts-method}
\alias{featureID<-,Paths-method}
\alias{featureID<-,SGFeatures-method}
\alias{from}
\alias{from,Paths-method}
\alias{from,SGVariantCounts-method}
\alias{from<-}
\alias{from<-,Paths-method}
\alias{from<-,SGVariantCounts-method}
\alias{geneID}
\alias{geneID,Counts-method}
\alias{geneID,Paths-method}
\alias{geneID,SGFeatures-method}
\alias{geneID<-}
\alias{geneID<-,Counts-method}
\alias{geneID<-,Paths-method}
\alias{geneID<-,SGFeatures-method}
\alias{geneName}
\alias{geneName,Counts-method}
\alias{geneName,Features-method}
\alias{geneName,Paths-method}
\alias{geneName<-}
\alias{geneName<-,Counts-method}
\alias{geneName<-,Features-method}
\alias{geneName<-,Paths-method}
\alias{segmentID}
\alias{segmentID,Paths-method}
\alias{segmentID,SGVariantCounts-method}
\alias{segmentID<-}
\alias{segmentID<-,Paths-method}
\alias{segmentID<-,SGVariantCounts-method}
\alias{slots}
\alias{splice3p}
\alias{splice3p,SGFeatureCounts-method}
\alias{splice3p,SGFeatures-method}
\alias{splice3p,SGSegments-method}
\alias{splice3p<-}
\alias{splice3p<-,SGFeatureCounts-method}
\alias{splice3p<-,SGFeatures-method}
\alias{splice3p<-,SGSegments-method}
\alias{splice5p}
\alias{splice5p,SGFeatureCounts-method}
\alias{splice5p,SGFeatures-method}
\alias{splice5p,SGSegments-method}
\alias{splice5p<-}
\alias{splice5p<-,SGFeatureCounts-method}
\alias{splice5p<-,SGFeatures-method}
\alias{splice5p<-,SGSegments-method}
\alias{to}
\alias{to,Paths-method}
\alias{to,SGVariantCounts-method}
\alias{to<-}
\alias{to<-,Paths-method}
\alias{to<-,SGVariantCounts-method}
\alias{txName}
\alias{txName,Counts-method}
\alias{txName,Features-method}
\alias{txName,Paths-method}
\alias{txName<-}
\alias{txName<-,Counts-method}
\alias{txName<-,Features-method}
\alias{txName<-,Paths-method}
\alias{type}
\alias{type,Counts-method}
\alias{type,Features-method}
\alias{type,Paths-method}
\alias{type<-}
\alias{type<-,Counts-method}
\alias{type<-,Features-method}
\alias{type<-,Paths-method}
\alias{variantID}
\alias{variantID,SGVariantCounts-method}
\alias{variantID,SGVariants-method}
\alias{variantID<-}
\alias{variantID<-,SGVariantCounts-method}
\alias{variantID<-,SGVariants-method}
\alias{variantName}
\alias{variantName,SGVariantCounts-method}
\alias{variantName,SGVariants-method}
\alias{variantName<-}
\alias{variantName<-,SGVariantCounts-method}
\alias{variantName<-,SGVariants-method}
\alias{variantType}
\alias{variantType,SGVariantCounts-method}
\alias{variantType,SGVariants-method}
\alias{variantType<-}
\alias{variantType<-,SGVariantCounts-method}
\alias{variantType<-,SGVariants-method}
\title{Accessing and replacing column slots}
\usage{
type(object)

type(object) <- value

txName(object)

txName(object) <- value

geneName(object)

geneName(object) <- value

featureID(object)

featureID(object) <- value

geneID(object)

geneID(object) <- value

splice5p(object)

splice5p(object) <- value

splice3p(object)

splice3p(object) <- value

from(object)

from(object) <- value

to(object)

to(object) <- value

segmentID(object)

segmentID(object) <- value

variantID(object)

variantID(object) <- value

eventID(object)

eventID(object) <- value

closed5p(object)

closed5p(object) <- value

closed3p(object)

closed3p(object) <- value

variantType(object)

variantType(object) <- value

variantName(object)

variantName(object) <- value

featureID5p(object)

featureID5p(object) <- value

featureID3p(object)

featureID3p(object) <- value

\S4method{type}{Features}(object)

\S4method{type}{Paths}(object)

\S4method{type}{Counts}(object)

\S4method{type}{Features}(object) <- value

\S4method{type}{Paths}(object) <- value

\S4method{type}{Counts}(object) <- value

\S4method{txName}{Features}(object)

\S4method{txName}{Paths}(object)

\S4method{txName}{Counts}(object)

\S4method{txName}{Features}(object) <- value

\S4method{txName}{Paths}(object) <- value

\S4method{txName}{Counts}(object) <- value

\S4method{geneName}{Features}(object)

\S4method{geneName}{Paths}(object)

\S4method{geneName}{Counts}(object)

\S4method{geneName}{Features}(object) <- value

\S4method{geneName}{Paths}(object) <- value

\S4method{geneName}{Counts}(object) <- value

\S4method{featureID}{SGFeatures}(object)

\S4method{featureID}{Paths}(object)

\S4method{featureID}{Counts}(object)

\S4method{featureID}{SGFeatures}(object) <- value

\S4method{featureID}{Paths}(object) <- value

\S4method{featureID}{Counts}(object) <- value

\S4method{geneID}{SGFeatures}(object)

\S4method{geneID}{Paths}(object)

\S4method{geneID}{Counts}(object)

\S4method{geneID}{SGFeatures}(object) <- value

\S4method{geneID}{Paths}(object) <- value

\S4method{geneID}{Counts}(object) <- value

\S4method{splice5p}{SGFeatures}(object)

\S4method{splice5p}{SGSegments}(object)

\S4method{splice5p}{SGFeatureCounts}(object)

\S4method{splice5p}{SGFeatures}(object) <- value

\S4method{splice5p}{SGSegments}(object) <- value

\S4method{splice5p}{SGFeatureCounts}(object) <- value

\S4method{splice3p}{SGFeatures}(object)

\S4method{splice3p}{SGSegments}(object)

\S4method{splice3p}{SGFeatureCounts}(object)

\S4method{splice3p}{SGFeatures}(object) <- value

\S4method{splice3p}{SGSegments}(object) <- value

\S4method{splice3p}{SGFeatureCounts}(object) <- value

\S4method{segmentID}{Paths}(object)

\S4method{segmentID}{SGVariantCounts}(object)

\S4method{segmentID}{Paths}(object) <- value

\S4method{segmentID}{SGVariantCounts}(object) <- value

\S4method{from}{Paths}(object)

\S4method{from}{SGVariantCounts}(object)

\S4method{from}{Paths}(object) <- value

\S4method{from}{SGVariantCounts}(object) <- value

\S4method{to}{Paths}(object)

\S4method{to}{SGVariantCounts}(object)

\S4method{to}{Paths}(object) <- value

\S4method{to}{SGVariantCounts}(object) <- value

\S4method{eventID}{SGVariants}(object)

\S4method{eventID}{SGVariantCounts}(object)

\S4method{eventID}{SGVariants}(object) <- value

\S4method{eventID}{SGVariantCounts}(object) <- value

\S4method{variantID}{SGVariants}(object)

\S4method{variantID}{SGVariantCounts}(object)

\S4method{variantID}{SGVariants}(object) <- value

\S4method{variantID}{SGVariantCounts}(object) <- value

\S4method{closed5p}{SGVariants}(object)

\S4method{closed5p}{SGVariantCounts}(object)

\S4method{closed5p}{SGVariants}(object) <- value

\S4method{closed5p}{SGVariantCounts}(object) <- value

\S4method{closed3p}{SGVariants}(object)

\S4method{closed3p}{SGVariantCounts}(object)

\S4method{closed3p}{SGVariants}(object) <- value

\S4method{closed3p}{SGVariantCounts}(object) <- value

\S4method{variantName}{SGVariants}(object)

\S4method{variantName}{SGVariantCounts}(object)

\S4method{variantName}{SGVariants}(object) <- value

\S4method{variantName}{SGVariantCounts}(object) <- value

\S4method{variantType}{SGVariants}(object)

\S4method{variantType}{SGVariantCounts}(object)

\S4method{variantType}{SGVariants}(object) <- value

\S4method{variantType}{SGVariantCounts}(object) <- value

\S4method{featureID5p}{SGVariants}(object)

\S4method{featureID5p}{SGVariantCounts}(object)

\S4method{featureID5p}{SGVariants}(object) <- value

\S4method{featureID5p}{SGVariantCounts}(object) <- value

\S4method{featureID3p}{SGVariants}(object)

\S4method{featureID3p}{SGVariantCounts}(object)

\S4method{featureID3p}{SGVariants}(object) <- value

\S4method{featureID3p}{SGVariantCounts}(object) <- value
}
\arguments{
\item{object}{Object containing column slot}

\item{value}{Replacement value}
}
\value{
Column value for accessor functions, updated object for
replacement functions.
}
\description{
Accessor and replacement functions for column slots.
}
\details{
S4 classes defined in the \code{SGSeq} package contain columns that
store information for each element in the object. For example, class
\code{TxFeatures} contains a column \code{type} that indicates feature
type. The specific columns contained in an object depend on its class.

To facilitate accessing and modifying columns, for each column
there exists a function with name identical to the column name
that can be used to access and modify it (see examples).
}
\examples{
head(type(txf))
head(type(sgf))
}
\author{
Leonard Goldstein
}

