% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TxSegments}
\alias{TxSegments}
\title{Constructor function for S4 class \code{TxSegments}}
\usage{
TxSegments(x)
}
\arguments{
\item{x}{\code{GRangesList} of \code{SGFeatures} with appropriate
outer elementMetadata columns}
}
\value{
A \code{TxSegments} object
}
\description{
Creates an instance of S4 class \code{TxSegments} for storing
transcript segments.
}
\author{
Leonard Goldstein
}
\keyword{internal}

